[css-conflicts-fix] fix css conflictx
