+++ src/components/CommentSection/styles/CommentSection.scss
@@ -7,6 +7,16 @@
   font-size: 1.5rem;
   text-transform: uppercase;
 }
+
+h3.ui.header.commentsHeader{
+  color: black !important;
+
+}
+
+.ui.comments .comment .actions{
+width:100%;
+}
+
 .commentsContainer {
   width: 100%;
   margin: 2% 7% 10% 6%;
