+++ src/components/SocialMediaIcons/SocialMediaIcons.scss
@@ -1,20 +1,22 @@
+@import '../../styles/partials/colors';
+@import '../../styles/partials/styles';
+
 .menu.SocialMediaContainer {
   background-color: transparent;
-
+  
   &.right {
-    margin-right: 7.5rem;
     font-size: 1.1rem;
   }
 
   .item.socialMediaIcons {
-    border-right: 1.3px solid #ffffff;
+    border-right: 1.3px solid $white;
     padding-top: 0;
     padding-bottom: 0;
     &:last-child {
       border-right: 0;
     }
     .socialMediaIcon {
-      color: #ffffff;
+      color: $white;
     }
 
     &.smallIconPadding {
@@ -28,4 +30,8 @@
       font-size: 1rem;
     }
   }
+}
+
+@media (max-width: 1024px) {
+ 
 }
