+++ src/components/sectionHeader/Titleheader.scss
@@ -11,7 +11,7 @@
   padding: 0;
 }
 
-p {
+p.titletext{
   background-color: $black;
   color: $white;
   font-family: Roboto;
