+++ src/constants/index.js
@@ -54,4 +54,3 @@ module.exports = {
   advertimage,
   socialMediaLoginIcons
 };
-
