+++ src/containers/Homecontainer/HomeContainer.scss
@@ -77,12 +77,13 @@
   padding-top: 0.7rem;
 }
 
-.column.Advert img.ui.image {
+img.ui.image.AdvertImg{
   margin: 0 auto;
   width: 70%;
   margin: 0 auto;
   width: 50%;
   margin-top: 8.8rem;
+  display: block;
 }
 
 .ui {
@@ -138,7 +139,7 @@
 
 .title,
 .header {
-  color: white !important;
+  color: white;
 }
 
 .ui.card.description {
@@ -244,59 +245,11 @@ p.articleDescription {
     width: 97% !important;
     margin: 0 auto;
   }
-  &.modal {
-    padding-top: 15%;
-    width: 100%;
-    height: 100%;
-    margin: 0;
-    box-shadow: none;
-    > .content {
-      display: block;
-      width: 70%;
-      font-size: 1rem;
-      line-height: 1.4;
-      padding: 1.5rem;
-      background: #fff;
-      margin: 0 auto;
-      padding-bottom: 0;
-    }
-  }
 }
 
 .actions {
   width: 70%;
   margin: 0 auto;
-}
-
-.ui.modal {
-  > .actions {
-    padding: 1rem 1rem;
-    text-align: right;
-    border-top-width: 0;
-    background-color: #fff;
-  }
-  input {
-    width: 20rem;
-    border-width: 0;
-  }
-}
-
-.dimmed.dimmable > .ui {
-  &.animating.dimmer,
-  &.visible.dimmer {
-    padding-top: 14.7%;
-    background-color: transparent;
-  }
-}
-
-.ui {
-  &.active.dimmer {
-    padding-top: 14.7%;
-    background-color: transparent;
-  }
-  &.modal button {
-    width: 15% !important;
-  }
 }
 
 .ui.divided.grid:not([class*='vertically divided']) > .column:not(.row),
