+++ src/pages/Login/Login.test.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import SignIn from '.';
+import Login from '.';
 
 describe('App', () => {
   it('should match snapshot', () => {
-    const wrapper = shallow(<SignIn />);
+    const wrapper = shallow(<Login />);
     expect(wrapper).toMatchSnapshot();
   });
 });
