+++ src/pages/Login/index.jsx
@@ -1,7 +1,17 @@
 import React from 'react';
+import ModalComponent from '../../components/FormContainer';
+import SignInForm from '../../components/SignInForm';
 
-export default function Login() {
+function SignIn() {
   return (
-    <div />
+    <ModalComponent
+      title="sign in"
+      baseText="Create an account."
+      link="Sign up"
+      formLink="signup">
+      <SignInForm />
+    </ModalComponent>
   );
 }
+
+export default SignIn;
