+++ src/routes/index.js
@@ -1,7 +1,6 @@
 import Home from '../pages/Home';
-import Login from '../pages/Login';
 import SignUp from '../pages/SignUp';
-import SignIn from '../pages/SignIn';
+import Login from '../pages/Login';
 import NotFound from '../pages/NotFound';
 import ViewArticle from '../pages/ViewArticle';
 
@@ -29,12 +28,6 @@ export default [
     component: SignUp,
     exact: true,
     path: '/signup'
-  },
-  {
-    name: 'signin',
-    component: SignIn,
-    exact: true,
-    path: '/signin'
   },
   {
     name: '404',
